+++ hackathon-den-strategy/src/interfaces/utils/integrations/IRebalancer.sol
@@ -3,10 +3,6 @@
 pragma solidity 0.8.26;
 
 interface IRebalancer {
-    function swap(
-        address sentCurrency,
-        uint sentAmount,
-        address receivedCurrency,
-        bytes calldata payload
-    ) external;
+    function swap(address sentCurrency, uint256 sentAmount, address receivedCurrency, bytes calldata payload)
+        external;
 }
